% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_helper.R
\name{get_data_partition}
\alias{get_data_partition}
\title{Partition data}
\usage{
get_data_partition(input_data, target_col, train_ratio, split_y = FALSE)
}
\arguments{
\item{input_data}{data.frame with all the data}

\item{target_col}{name or column number of a target variable}

\item{train_ratio}{what should be the ratio}

\item{split_y}{should the target variable be returned separately?}
}
\value{

}
\description{
This function divide data into training and holdout set and returns them as a list.
It uses caret package createDataPartition() function to make the division.
}
\examples{
\dontrun{
.
}
}
